refactor(domain): Round 확인에 대한 책임을 Game에서 RoundCount로 이동
Game은 RoundCount에게 전체 라운드 횟수를 전달 받아 직접 계산하고 있었다. 하지만 이런 방식은 RoundCount의 내부 상태를 전부 오픈하게 되고 Game이 너무 많은 책임을 맡게된다. 때문에 Round 관련 로직은 Round가 맡아서 캡슐화하는 것이 더 좋은 방법이라고 생각된다.

https://emadam.notion.site/refactor-aea8071f8da4477181a43899475d1cb6
