+++ src/main/java/racingcar/domain/RoundCount.java
@@ -5,14 +5,21 @@ public class RoundCount {
     public static final int MIN_COUNT = 0;
 
     private final int count;
+    private int current = 0;
 
     public RoundCount(String count) {
         validate(count);
         this.count = Integer.parseInt(count);
     }
 
-    public int getCount() {
-        return count;
+    public void nextRound() {
+        validateRound();
+
+        current++;
+    }
+
+    public boolean hasNextRound() {
+        return (current < count);
     }
 
     private void validate(String count) {
@@ -39,5 +46,12 @@ public class RoundCount {
 
     private boolean isNegativeNumber(int count) {
         return (count < MIN_COUNT);
+    }
+
+    private void validateRound() {
+        if (!hasNextRound()) {
+            throw new IllegalStateException(
+                    ErrorMessage.START_ROUND_WHEN_GAME_OVER_ERROR.getMessage());
+        }
     }
 }
